requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.2.2
pytest-xdist==3.6.1
requests==2.32.3
selenium==4.22.0
webdriver-manager==4.0.1

